refactor methods into utils
