+++ refactor/dispatcher/app/core/utils.py
@@ -22,6 +22,7 @@
 
 from datetime import datetime, timezone
 from io import BytesIO
+from multiprocessing import Queue as MPQ
 from typing import Any, Dict, List
 
 import cloudpickle as pickle
@@ -41,6 +42,15 @@ from covalent._shared_files.defaults import (
     subscript_prefix,
 )
 from covalent._workflow.lattice import Lattice
+
+
+def is_empty(mp_queue: MPQ):
+    if elem := mp_queue.get():
+        mp_queue.put(elem)
+        return True
+    else:
+        mp_queue.put(None)
+        return False
 
 
 def preprocess_transport_graph(task_id: int, task_name: str, result_obj: Result) -> Result:
